+++ core/domain/CMakeLists.txt
@@ -4,5 +4,6 @@ project (domain)
 add_subdirectory (userlogin)
 
 if (BUILD_UNITTEST)
+    include_directories (.)
     add_subdirectory (unittest)
 endif()
