+++ scripts/run_cppcheck.sh
@@ -5,4 +5,4 @@ cd $SCRIPT_DIR/../
 # Note: When updating the line below, update .travis.yml as well
 cppcheck --std=c++11 --enable=warning,style,performance,portability,information \
  --suppress=missingIncludeSystem --error-exitcode=1 --inline-suppr \
- application core datamanager mock utility -icore/domain/unittest
+ core orchestra mock utility -icore/domain/unittest
