+++ scripts/run_cpplint.sh
@@ -4,4 +4,4 @@ cd $SCRIPT_DIR/../
 
 # Note: When updating the line below, update .travis.yml as well
 cpplint --filter=-readability/multiline_comment,-whitespace/ending_newline,-build/c++11 --linelength=100 \
---recursive application core datamanager mock utility
+--recursive orchestra core utility
